+++ app/services/contact/get_suggestions.rb
@@ -1,4 +1,6 @@
 class Contact::GetSuggestions
+  RETURN_CONTACTS_COUNT = 100
+
   attr_reader :owner_mkey
 
   def initialize(owner_mkey)
@@ -13,6 +15,6 @@ class Contact::GetSuggestions
   private
 
   def contacts
-    Contact.by_owner(owner_mkey)
+    Contact.by_owner(owner_mkey).select { |c| !c.marked_as_friend? }.take(RETURN_CONTACTS_COUNT)
   end
 end