+++ app/workers/notification/fake_user_joined_worker.rb
@@ -4,7 +4,7 @@ class Notification::FakeUserJoinedWorker
   def self.perform
     WriteLog.info(self, 'cron job was executed')
     Owner.subscribed.each do |owner|
-      contact = owner.not_proposed_contacts.first
+      contact = owner.contacts.not_proposed.first
       contact && Notification::Create.new(:fake_user_joined, contact).do.each { |n| n.send_notification }
     end
   end