+++ spec/models/owner_spec.rb
@@ -23,9 +23,9 @@ RSpec.describe Owner, type: :model do
   end
 
   describe '#not_proposed_contacts' do
-    subject { instance.not_proposed_contacts }
+    subject { instance.contacts.not_proposed }
     before do
-      FactoryGirl.create :notification, contact: contact_2
+      FactoryGirl.create(:notification, contact: contact_2)
     end
 
     it { is_expected.to eq [contact_1] }
@@ -36,8 +36,8 @@ RSpec.describe Owner, type: :model do
 
     context 'when unsubscribed' do
       before do
-        FactoryGirl.create :notification, contact: contact_1, status: 'added'
-        FactoryGirl.create :notification, contact: contact_2, status: 'unsubscribed'
+        FactoryGirl.create(:notification, contact: contact_1, status: 'added')
+        FactoryGirl.create(:notification, contact: contact_2, status: 'unsubscribed')
       end
 
       it { is_expected.to be true }
@@ -45,8 +45,8 @@ RSpec.describe Owner, type: :model do
 
     context 'when not unsubscribed' do
       before do
-        FactoryGirl.create :notification, contact: contact_1, status: 'ignored'
-        FactoryGirl.create :notification, contact: contact_2, status: 'added'
+        FactoryGirl.create(:notification, contact: contact_1, status: 'ignored')
+        FactoryGirl.create(:notification, contact: contact_2, status: 'added')
       end
 
       it { is_expected.to be false }